Update style.css
This update was focused on the "Get To The Mountain" button. You will notice I did some minor update to the design of the navigation. 

I was going for a more skeuomorphic design for the button, but also implement some flat button design.  I changed the font of the button to Raleway so it flows with the rest of the font. I added a black box shadow. When the user hovers the cursor over the button, the background and color will change and the box shadow will increase; giving it a pop-up effect. When the user presses the button, the button has a press-down effect by removing the box shadow and apply a very short delay.

I want to note, in order to achieve the pop up effect I had to edit the original HTML code (please see index.html for changes). Therefore, I had to add a few CSS attributes in order to re-center the button. These attributes include display, text-align, margin, and width.