Update style.css
In this update I added both a hover and active selectors to the .activity class. When a user hovers their cursor over the buttons, the background color will lighten up slightly and jump up. Once the user clicks on the button both the color and background color will change and a border will appear letting the user know they have pressed that button.